+++ STEPViewer/ProductDefinition.cpp
@@ -253,13 +253,11 @@ int32_t CProductDefinition::GetNextInstance()
 
 void CProductDefinition::Calculate()
 {
-	//#perf
 	setSTEPFormatSettings();
 
 	SdaiModel iSdaiModel = sdaiGetInstanceModel(GetInstance());
 	ASSERT(iSdaiModel != 0);
 
-	//#perf
 	setSegmentation(iSdaiModel, 16, 0.);
 
 	ASSERT(m_pVertexBuffer == nullptr);
@@ -277,8 +275,8 @@ void CProductDefinition::Calculate()
 	/**
 	* Retrieves the vertices
 	*/
-	m_pVertexBuffer->data() = new float[(uint32_t)m_pVertexBuffer->size() * m_pVertexBuffer->getVertexLength()];
-	memset(m_pVertexBuffer->data(), 0, (uint32_t)m_pVertexBuffer->size() * m_pVertexBuffer->getVertexLength() * sizeof(float));
+	m_pVertexBuffer->data() = new float[(uint32_t)m_pVertexBuffer->size() * (int64_t)m_pVertexBuffer->getVertexLength()];
+	memset(m_pVertexBuffer->data(), 0, (uint32_t)m_pVertexBuffer->size() * (int64_t)m_pVertexBuffer->getVertexLength() * sizeof(float));
 
 	UpdateInstanceVertexBuffer(m_iInstance, m_pVertexBuffer->data());
